% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModelL1.R
\name{LinearModelL1}
\alias{LinearModelL1}
\title{Linear Model algorithm with L1 regularization}
\usage{
LinearModelL1(X.scaled.mat, y.vec, penalty, opt.thresh, initial.weight.vec,
  step.size)
}
\arguments{
\item{step.size}{}
}
\description{
This algorithm takes one penalty value
}
