% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModelL1penalties.R
\name{LinearModelL1penalties}
\alias{LinearModelL1penalties}
\title{Linear Model with L1 regularization}
\usage{
LinearModelL1penalties(X.mat, y.vec, penalty.vec = seq(0.3, 0, by =
  -0.03), step.size = 0.01)
}
\arguments{
\item{step.size}{}
}
\description{
This algorithm takes the penalty vector as input and iterate through each value in the vector using LinearModelL1
}
