% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMIterations.R
\name{LMLogisticLossIterations}
\alias{LMLogisticLossIterations}
\title{Linear model iteration with logistic loss}
\usage{
LMLogisticLossIterations(X.mat, y.vec, max.iterations, step.size)
}
\arguments{
\item{X.mat}{train feature matrix of size [n x p]}

\item{y.vec}{train label vector of size [n x 1]}

\item{max.iterations}{integer scalar greater than 1}

\item{step.size}{a numeric scalar greater than zero}
}
\value{
W.mat matrix of weight vectors of size [(p + 1) x max.iterations]
}
\description{
Training by using gradient descent on a linear model with logistic loss .
Return a matrix of weight vector for each iteration to the max iteration.
}
\examples{

data(spam, package = "ElemStatLearn")
X.mat <- as.matrix(spam[, 1:57])
y.vec <- ifelse(spam$spam == "spam", 1, -1)
W.mat <- LMLogisticLossIterations(X.mat, y.vec, max.iterations = 100L, step.size = 0.5)
(W.mat)
}
