% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LML2penalties.R
\name{LMLogisticLossL2penalties}
\alias{LMLogisticLossL2penalties}
\title{Linear model with L2 penalties and logistic loss}
\usage{
LMLogisticLossL2penalties(X.mat, y.vec, penalty.vec, opt.thresh = 0.5)
}
\arguments{
\item{X.mat}{a numeric matrix of size [n x p]}

\item{y.vec}{a numeric vector of length nrow(X.mat)}

\item{penalty.vec}{a non-negative numeric vector}
}
\value{
W.mat a numeric weight matrix of size [ncol(X.mat) x length(penalty.vec)]
}
\description{
Training by using L2 regularization on a linear model with logistic loss .
Return a matrix of weight vector for the given penalty vector.
}
\examples{
data(spam, package = "ElemStatLearn")
X.mat <- as.matrix(spam[, 1:57])
y.vec <- ifelse(spam$spam == "spam", 1, -1)
penalty.vec <- seq(5:0.1, by = -0.1)
W.mat <- LMLogisticLossL2penalties(X.mat, y.vec, penalty.vec, 0.5)
(W.mat)
}
