% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMIterations.R
\name{LMSquareLossIterations}
\alias{LMSquareLossIterations}
\title{Linear model iteration with square loss}
\usage{
LMSquareLossIterations(X.mat, y.vec, max.iterations, step.size = 0.5)
}
\arguments{
\item{X.mat}{train feature matrix of size [n x p]}

\item{y.vec}{train label vector of size [n x 1]}

\item{max.iterations}{integer scalar greater than 1}

\item{step.size}{integer scalar}
}
\value{
W.mat matrix of weight vectors of size [(p + 1) x max.iterations].
A prediction can be obtained by cbind(1,X.mat) %*% W.mat.
}
\description{
Linear model iteration with square loss
}
