% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LML2.R
\name{LMSquareLossL2}
\alias{LMSquareLossL2}
\title{Linear model L2 regularization with square loss}
\usage{
LMSquareLossL2(X.scaled.mat, y.vec, penalty, opt.thresh = 0.5,
  initial.weight.vec)
}
\arguments{
\item{X.scaled.mat}{a numeric matrix of size [n x p]}

\item{y.vec}{a numeric matrix of length nrow(X.scaled.mat)}

\item{penalty}{a non-negative numeric scalar}

\item{opt.thresh}{a positive numeric scalar}

\item{initial.weight.vec}{a numeric vector of size ncol(X.scaled.mat)}
}
\value{
opt.weight the optimal weight vector of length ncol(X.scaled)
}
\description{
Linear model L2 regularization with square loss
}
