% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LML2CV.R
\name{LMSquareLossL2CV}
\alias{LMSquareLossL2CV}
\title{Cross validation using linear model with L2 regularization and square loss}
\usage{
LMSquareLossL2CV(X.mat, y.vec, fold.vec, penalty.vec)
}
\arguments{
\item{X.mat}{a numeric matrix of size [n x p]}

\item{y.vec}{a numeric vector of length nrow(X.mat)}

\item{fold.vec}{a numeric vector of length nrow(X.mat)}

\item{penalty.vec}{a non-negative numeric vector}
}
\value{
result.list a list with mean.validation.loss.vec,
mean.train.loss.vec,penalty.vec,selected.penalty,weight.vec,and predict function
}
\description{
Cross validation using linear model with L2 regularization and square loss
}
