% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LML2penalties.R
\name{LMSquareLossL2penalties}
\alias{LMSquareLossL2penalties}
\title{Linear model with L2 penalties and square loss}
\usage{
LMSquareLossL2penalties(X.mat, y.vec, penalty.vec, opt.thresh = 0.5)
}
\arguments{
\item{X.mat}{a numeric matrix of size [n x p]}

\item{y.vec}{a numeric vector of length nrow(X.mat)}

\item{penalty.vec}{a non-negative numeric vector}

\item{opt.thresh}{a non-negative numeric scalar}
}
\value{
W.mat a numeric weight matrix of size [ncol(X.mat) x length(penalty.vec)]
}
\description{
Training by using L2 regularization on a linear model with square loss .
Return a matrix of weight vector for the given penalty vector.
}
\examples{
data(ozone, package = "ElemStatLearn")
y.vec <- ozone[, 1]
X.mat <- as.matrix(ozone[,-1])
W.mat <- LMSquareLossL2penalties(X.mat, y.vec,seq(5, 0.1, by = -0.1))

}
