% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN1toKmaxPredict.R
\name{NN1toKmaxPredict}
\alias{NN1toKmaxPredict}
\title{1 to k nearest neighbor algorithm}
\usage{
NN1toKmaxPredict(X.mat, y.vec, testX.mat, max.neighbors)
}
\arguments{
\item{X.mat}{numeric train feature matrix [n x p]}

\item{y.vec}{numeric train label vector [n x 1], 0/1 for binary classification, real number for regression.}

\item{textX.mat}{numeric test feature matrix}

\item{max.neighbor}{scalar integer, max number of neighbor}
}
\description{
R function that wraps the C++ code.
}
