% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNLearnCV.R
\name{NNLearnCV}
\alias{NNLearnCV}
\title{Cross-validation using nearest neighbors}
\usage{
NNLearnCV(X.mat, y.vec, max.neighbors = 30L, fold.vec = NULL,
  n.folds = 5L)
}
\arguments{
\item{X.mat}{numeric train feature matrix [n x p]}

\item{y.vec}{numeric train label vector [n x 1], 0/1 for binary classification, real number for regression.}

\item{max.neighbors}{scalar integer, max number of the neighbors}

\item{fold.vec}{integer vector that holds fold ID number [n x 1]}

\item{n.folds}{scalar integer, number of the folds}
}
\description{
Using nearest neighbors algorithm for cross validation to find out the best K and give a prediction of the test data based on it.
}
