% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEarlyStoppingCV.R
\name{LMLogisticLossEarlyStoppingCV}
\alias{LMLogisticLossEarlyStoppingCV}
\title{Cross validation algorithm using linear model with logistic loss}
\usage{
LMLogisticLossEarlyStoppingCV(X.mat, y.vec, fold.vec = NULL,
  max.iteration, step.size = 0.5)
}
\arguments{
\item{X.mat}{train feature matrix of size [n x p]}

\item{y.vec}{train label vector of size [n x 1]}

\item{fold.vec}{fold index vector of size [n x 1]}

\item{max.iteration}{integer scalar greater than 1}

\item{step.size}{a numeric scaler greater than 0, default is 0.5}
}
\value{
result.list a list with mean.validation.loss.vec,
mean.train.loss.vec,selected.steps,weight.vec,and predict function
}
\description{
Cross validation algorithm using linear model with logistic loss
}
\examples{

data(spam, package = "ElemStatLearn")
X.mat <- as.matrix(spam[, 1:57])
y.vec <- ifelse(spam$spam == "spam", 1, 0)
earlystopping.list <- LMLogisticLossEarlyStoppingCV(X.mat, y.vec, NULL, m100L, 0.5)
(earlystopping.list$predict(X.mat[1,]))

}
