% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMIterations.R
\name{LMLogisticLossIterations}
\alias{LMLogisticLossIterations}
\title{Linear model iteration with logistic loss}
\usage{
LMLogisticLossIterations(X.mat, y.vec, max.iterations, step.size)
}
\arguments{
\item{X.mat}{train feature matrix of size [n x p]}

\item{y.vec}{train label vector of size [n x 1]}

\item{max.iterations}{integer scalar greater than 1}

\item{step.size}{integer scalar}
}
\value{
W.mat matrix of weight vectors of size [n x max.iterations]
}
\description{
Linear model iteration with logistic loss
}
