% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LML2.R
\name{LMLogisticLossL2}
\alias{LMLogisticLossL2}
\title{Linear model L2 regularization with logistic loss, including beta during the training}
\usage{
LMLogisticLossL2(X.scaled.mat, y.vec, penalty, opt.thresh,
  initial.weight.vec, step.size = 0.01, max.iteration = 10)
}
\arguments{
\item{X.scaled.mat}{a numeric matrix of size [n x p]}

\item{y.vec}{a numeric matrix of length nrow(X.scaled.mat)}

\item{penalty}{a non-negative numeric scalar}

\item{opt.thresh}{a positive numeric scalar}

\item{initial.weight.vec}{a numeric vector of size ncol(X.scaled.mat)}

\item{step.size}{a numeric scalar greater than zero}

\item{max.iteration}{a integer scalar greater than one}
}
\value{
opt.weight the optimal weight vector of length ncol(X.scaled)
}
\description{
Training by using L2 regularization on a linear model with logistic loss .
Return the optimal weight vector for the given threshold and penalty.
}
\examples{

data(spam, package = "ElemStatLearn")
X.mat <- as.matrix(spam[, 1:57])
X.scaled.mat <- scale(X.mat)
y.vec <- ifelse(spam$spam == "spam", 1, -1)
opt.weight.vec <- LMLogisticLossL2(X.mat, y.vec, 0.5, 0.5, rep(0,ncol(X.scaled.mat), 0.01, 100L))
(opt.weigh.vec)
}
