% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LML2penalties.R
\name{LMLogisticLossL2penalties}
\alias{LMLogisticLossL2penalties}
\title{Linear model with L2 penalties and logistic loss}
\usage{
LMLogisticLossL2penalties(X.mat, y.vec, penalty.vec, opt.thresh = 5)
}
\arguments{
\item{X.mat}{a numeric matrix of size [n x p]}

\item{y.vec}{a numeric vector of length nrow(X.mat)}

\item{penalty.vec}{a non-negative numeric vector}
}
\value{
W.mat a numeric weight matrix of size [ncol(X.mat) x length(penalty.vec)]
}
\description{
Linear model with L2 penalties and logistic loss
}
