% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEarlyStoppingCV.R
\name{LMSquareLossEarlyStoppingCV}
\alias{LMSquareLossEarlyStoppingCV}
\title{Cross validation algorithm using linear model with square loss}
\usage{
LMSquareLossEarlyStoppingCV(X.mat, y.vec, fold.vec, max.iteration)
}
\arguments{
\item{X.mat}{train feature matrix of size [n x p]}

\item{y.vec}{train label vector of size [n x 1]}

\item{max.iteration}{integer scalar greater than 1}
}
\value{
result.list a list with mean.validation.loss.vec,
mean.train.loss.vec,selected.steps,weight.vec,and predict function
}
\description{
Cross validation algorithm using linear model with square loss
}
