% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEarlyStoppingCV.R
\name{LMSquareLossEarlyStoppingCV}
\alias{LMSquareLossEarlyStoppingCV}
\title{Cross validation algorithm using linear model with square loss}
\usage{
LMSquareLossEarlyStoppingCV(X.mat, y.vec, fold.vec, max.iteration)
}
\arguments{
\item{X.mat}{train feature matrix of size [n x p]}

\item{y.vec}{train label vector of size [n x 1]}

\item{fold.vec}{the fold vector information of size [n x 1], which is assigned in the general test}

\item{max.iteration}{integer scalar greater than 1}
}
\value{
result.list a list with mean.validation.loss.vec,
mean.train.loss.vec,selected.steps,weight.vec,and predict function
}
\description{
Training by using cross validation on a linear model with square loss and early stopping method.
Return a list which contains the best iteration step, mean loss of training and validation data,
and a predict function which gives a prediction based on the selected step.
}
\examples{
data(ozone, package = "ElemStatLearn")
y.vec <- ozone[, 1]
X.mat <- as.matrix(ozone[,-1])
result.list <- LMSquareLossEarlyStoppingCV(X.mat, y.vec, max.iteration = 5L)
result.lsit$predict(X.mat[c(3,19,20),])
}
