% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LML2CV.R
\name{LMSquareLossL2CV}
\alias{LMSquareLossL2CV}
\title{Cross validation using linear model with L2 regularization and square loss}
\usage{
LMSquareLossL2CV(X.mat, y.vec, fold.vec, penalty.vec)
}
\arguments{
\item{X.mat}{a numeric matrix of size [n x p]}

\item{y.vec}{a numeric vector of length nrow(X.mat)}

\item{fold.vec}{a numeric vector of length nrow(X.mat)}

\item{penalty.vec}{a non-negative numeric vector}
}
\value{
result.list a list with mean.validation.loss.vec,
mean.train.loss.vec,penalty.vec,selected.penalty,weight.vec,and predict function
}
\description{
Training by using cross validation on a linear model with square loss and L2 regularization.
Return a list which contains the best penalty, mean loss of training and validation data,
and a predict function which gives a prediction based on the selected penalty.
}
\examples{
data(ozone, package = "ElemStatLearn")
y.vec <- ozone[, 1]
X.mat <- as.matrix(ozone[,-1])
result.list <- LMSquareLossL2CV(X.mat, y.vec, penalty.vec = seq(5, 0.1, by = -0.1))
result.list$predict(X.mat[c(3,4,11),])
}
